Add Parameters column to Classes table
This commit adds a Parameter column to the Classes table that is listed
in Node Group and Node pages.  The text 'Source is not #{resource.name}'
is shows if the class is not a member of the node or node group being
viewed.  If the class is a member, the link Show Parameters is displayed
and takes the user to the parameters for that class membership