(RE-453) Make a 2.8.x branch for pe-puppet-dashboard
This brings over all of the goodness of pe-rpm-puppet-dashboard and
pe-deb-puppet-dashboard and also has some fun bugfixes. This allows
pe-puppet-dashboard to build using the other PE project packaging pipeline
tools. It moves any shared packaging files like logrotate, cronjobs, apache
auth rules, into ext so that there is one source of truth. It also applies two
patches that were being carried and removes the patch files. It also adds some
logic to the rakefile so that rails is not started up if a packaging task is
invoked, as that causes serious problems on new versions of ruby/gems.  These
builds have been verified to be file equivalent to the 2.8.3 packages, except
for the addition of some files in /opt/puppet/share/puppet-dashboard/ext
(shared packaging files).